Warn user if interface cannot be detected
We also tell the user how this situation can be fixed (by asking them to
send us the output if `ifconfig` or `ipconfig`).  This is much more
helpful than "uninitialized value" warnings.

If the interface can't be detected, we return the empty string from the
`ip()` sub so that *something* turns up, but not something that's going
to be misleading.